bug fixed on desktop app

+++ PracticeWebApps/PracticeWebApps_LogicLibrary/Managers/FileManager.cs
@@ -1,24 +1,18 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PracticeWebApps_LogicLibrary.Managers
+﻿namespace PracticeWebApps_LogicLibrary.Managers
 {
     public class FileManager
     {
-        string originalFileName;
+        private string originalFileName;
         public string SaveFile(string selectedFilePath)
         {
-            if (string.IsNullOrEmpty(selectedFilePath) && File.Exists(selectedFilePath))
+            if (string.IsNullOrEmpty(selectedFilePath) || !File.Exists(selectedFilePath))
             {
                 throw new ArgumentException("Select a file before saving");
             }
 
             try
             {
-                string wwwrootPath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot");
+                string wwwrootPath = "C:\\Sem2-project-repository\\individual-project-nazim-ahmedov-s2\\PracticeWebApps\\PracticeWebApps\\wwwroot\\";
 
                 string destinationFolder = Path.Combine(wwwrootPath, "images");
 
@@ -31,33 +25,31 @@ namespace PracticeWebApps_LogicLibrary.Managers
 
                 string destinationPath = Path.Combine(destinationFolder, originalFileName);
 
-                File.Copy(selectedFilePath, destinationPath, true);
+                if (!File.Exists(destinationPath))
+                {
+                    File.Copy(selectedFilePath, destinationPath, true);
+                }
 
                 return destinationPath;
             }
             catch (ArgumentException ex)
             {
-                throw new ArgumentException($"An error occured: {ex.Message}");
+                throw new ArgumentException($"An error occured: {ex}");
             }
             
         }
-        public string GetFileName()
-        {
-            return originalFileName;
-        }
-        public string CreateRelativePath(string originalFileName, string subfolder)
+        public string CreateRelativePath(string subfolder)
         {
             try
             {
-
-                string relativePath = Path.Combine(subfolder, originalFileName).Replace('\\', '/'); ;
+                string relativePath = Path.Combine(subfolder, originalFileName).Replace('\\', '/');
 
                 return $"~/{relativePath}";
 
             }
             catch (ArgumentException ex)
             {
-                throw new ArgumentException($"An error occured: {ex.Message}");
+                throw new ArgumentException($"An error occured: {ex}");
             }
         }
     }
